id manager update: issn check
issn manager was not updated following the same procedure as for the other id managers since it has no API and thus ids can just be normalised. Therefore, there is no need to store validated ids in a more efficient way. The ISSN tests were stored in a separate file, as for the other identifier managers