id manager check: ISBN
isbn manager was not updated following the same procedure as for the other id managers since it has no API and thus ids can just be normalised. Therefore, there is no need to store validated ids in a more efficient way. The ISBN tests were stored in a separate file, as for the other identifier managers